ref(onboarding): Split tornado onboarding docs (#103035)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
